Raspberry Pi Image Builder
Depends: kpartx
Added modified version of the PiShrink script.
https://github.com/Drewsif/PiShrink

Need to figure out how to accomplish this using btrfs.
Do not try to build a btrfs image in this branch or it will fail.

The goal is to remove the need for set image sizes. This will give
us more options, such as adding Desktop possibilities on the fly.