README: Going forward
I will no longer be releasing images. The time and effort it takes to continue this is just not acceptable and the reward for doing so is nothing. I will still continue to support kernels using fetch and the kernel helper script in the builder, but beyond that I'm out.